shooting mobs

+++ src/com/thecherno/rain/entity/Entity.java
@@ -18,7 +18,7 @@ public class Entity {
     }
 
     public void render(Screen screen) {
-        if(sprite != null) screen.renderSprite((int) x, (int) y, sprite, true);
+        if(sprite != null) screen.renderSprite( x, y, sprite, true);
     }
 
     public void remove() {
